modularized the D3 component by separating the packchart component from the D3 component
