+++ package.json
@@ -44,6 +44,7 @@
     "react-dom": "^18.2.0",
     "react-router": "^6.22.2",
     "react-router-dom": "^6.22.2",
+    "reactflow": "^11.10.4",
     "style-loader": "^3.3.4",
     "webpack-cli": "^5.1.4"
   },
